requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
Faker>=8.12,<9.0
